+++ src/mcp/utils.ts
@@ -5,7 +5,7 @@ import type { TaskStore } from '@modelcontextprotocol/sdk/experimental/tasks/int
 import type { ApifyClient } from 'apify-client';
 
 import { processInput } from '../input.js';
-import type { ActorStore, Input, UiMode } from '../types.js';
+import type { ActorStore, Input, ServerMode } from '../types.js';
 import { loadToolsFromInput } from '../utils/tools_loader.js';
 import { MAX_TOOL_NAME_LENGTH, SERVER_ID_LENGTH } from './const.js';
 
@@ -41,11 +41,11 @@ export function getProxyMCPServerToolName(url: string, toolName: string): string
  * If URL contains query parameter `actors`, return tools from Actors otherwise return null.
  * @param url The URL to process
  * @param apifyClient The Apify client instance
- * @param uiMode UI mode from server options
+ * @param mode Server mode for tool variant resolution
  */
-export async function processParamsGetTools(url: string, apifyClient: ApifyClient, uiMode?: UiMode, actorStore?: ActorStore) {
+export async function processParamsGetTools(url: string, apifyClient: ApifyClient, mode: ServerMode, actorStore?: ActorStore) {
     const input = parseInputParamsFromUrl(url);
-    return await loadToolsFromInput(input, apifyClient, uiMode, actorStore);
+    return await loadToolsFromInput(input, apifyClient, mode, actorStore);
 }
 
 export function parseInputParamsFromUrl(url: string): Input {
