+++ src/tools/common/abort_actor_run.ts
@@ -1,0 +1,52 @@
+import { z } from 'zod';
+
+import { createApifyClientWithSkyfireSupport } from '../../apify_client.js';
+import { HelperTools } from '../../const.js';
+import type { InternalToolArgs, ToolEntry, ToolInputSchema } from '../../types.js';
+import { compileSchema } from '../../utils/ajv.js';
+
+const abortRunArgs = z.object({
+    runId: z.string()
+        .min(1)
+        .describe('The ID of the Actor run to abort.'),
+    gracefully: z.boolean().optional().describe('If true, the Actor run will abort gracefully with a 30-second timeout.'),
+});
+
+/**
+ * https://docs.apify.com/api/v2/actor-run-abort-post
+ */
+export const abortActorRun: ToolEntry = Object.freeze({
+    type: 'internal',
+    name: HelperTools.ACTOR_RUNS_ABORT,
+    description: `Abort an Actor run that is currently starting or running.
+For runs with status FINISHED, FAILED, ABORTING, or TIMED-OUT, this call has no effect.
+The results will include the updated run details after the abort request.
+
+USAGE:
+- Use when you need to stop a run that is taking too long or misconfigured.
+
+USAGE EXAMPLES:
+- user_input: Abort run y2h7sK3Wc
+- user_input: Gracefully abort run y2h7sK3Wc`,
+    inputSchema: z.toJSONSchema(abortRunArgs) as ToolInputSchema,
+    /**
+     * Allow additional properties for Skyfire mode to pass `skyfire-pay-id`.
+     */
+    ajvValidate: compileSchema({ ...z.toJSONSchema(abortRunArgs), additionalProperties: true }),
+    requiresSkyfirePayId: true,
+    annotations: {
+        title: 'Abort Actor run',
+        readOnlyHint: false,
+        destructiveHint: true,
+        idempotentHint: true,
+        openWorldHint: false,
+    },
+    call: async (toolArgs: InternalToolArgs) => {
+        const { args, apifyToken, apifyMcpServer } = toolArgs;
+        const parsed = abortRunArgs.parse(args);
+
+        const client = createApifyClientWithSkyfireSupport(apifyMcpServer, args, apifyToken);
+        const v = await client.run(parsed.runId).abort({ gracefully: parsed.gracefully });
+        return { content: [{ type: 'text', text: `\`\`\`json\n${JSON.stringify(v)}\n\`\`\`` }] };
+    },
+} as const);
