+++ src/tools/common/add_actor.ts
