+++ src/types.ts
@@ -94,7 +94,7 @@ export type ToolBase = z.infer<typeof ToolSchema> & {
     /**
      * Whether this tool is only available in OpenAI UI mode.
      * @deprecated No longer used for filtering. Mode-specific tools are resolved at build time
-     * via buildCategories(uiMode). Will be removed in a future release.
+     * via getCategoryTools(mode). Will be removed in a future release.
      */
     openaiOnly?: boolean;
 };
@@ -391,9 +391,35 @@ export type ToolCallTelemetryProperties = {
 };
 
 /**
- * UI mode for tool responses.
+ * Server mode controls which tool variants are served.
+ *
+ * - `'default'` — standard MCP tools for generic clients
+ * - `'openai'` — OpenAI-specific tool variants (async execution, widget metadata, internal tools)
+ *
+ * Every call site that resolves tool categories must pass an explicit ServerMode value.
+ * This prevents accidentally serving wrong-mode tools.
  */
-export type UiMode = 'openai';
+export type ServerMode = 'default' | 'openai';
+
+/** All valid server modes, for iteration in tests and caches. */
+export const SERVER_MODES: readonly ServerMode[] = ['default', 'openai'] as const;
+
+/**
+ * UI mode for tool responses — the external-facing subset of ServerMode.
+ * Derived from ServerMode: excludes 'default' since the absence of a UI mode means default.
+ */
+export type UiMode = Exclude<ServerMode, 'default'>;
+
+/** Set of valid UiMode values for O(1) membership checks at runtime. */
+const UI_MODES: ReadonlySet<string> = new Set<string>(SERVER_MODES.filter((m): m is UiMode => m !== 'default'));
+
+/**
+ * Parse an untrusted string into a valid UiMode, returning `undefined` for invalid values.
+ * Use at ingestion boundaries (URL params, env vars) to prevent invalid modes from propagating.
+ */
+export function parseUiMode(value: string | null | undefined): UiMode | undefined {
+    return value && UI_MODES.has(value) ? (value as UiMode) : undefined;
+}
 
 /**
  * Parameters for executing a direct actor tool (`type: 'actor'`).
@@ -428,7 +454,7 @@ export type ActorExecutionResult = {
 
 /**
  * Executor for direct actor tools (`type: 'actor'`).
- * Selected at server construction time based on uiMode.
+ * Selected at server construction time based on serverMode.
  * Default mode runs synchronously; OpenAI mode runs async with widget metadata.
  */
 export type ActorExecutor = {
@@ -534,7 +560,8 @@ export type ActorsMcpServerOptions = {
     /**
      * UI mode for tool responses.
      * - 'openai': OpenAI specific widget rendering
-     * If not specified, there will be no widget rendering.
+     * If not specified, defaults to 'default' mode (no widget rendering).
+     * Normalized to {@link ServerMode} at server construction.
      */
     uiMode?: UiMode;
 }
