+++ src/utils/server-instructions/index.ts
@@ -1,20 +1,21 @@
 /**
  * Server instructions entry point.
- * Selects the appropriate instructions based on UI mode.
+ * Selects the appropriate instructions based on server mode.
  */
 
-import type { UiMode } from '../../types.js';
+import type { ServerMode } from '../../types.js';
 import { getDefaultInstructions } from './default.js';
 import { getOpenaiInstructions } from './openai.js';
 
+/** Mode → instructions builder. Add new modes here. */
+const instructionsByMode: Record<ServerMode, () => string> = {
+    default: getDefaultInstructions,
+    openai: getOpenaiInstructions,
+};
+
 /**
- * Build server instructions for the given UI mode.
- *
- * @param uiMode - The UI mode ('openai' or undefined for default)
- * @returns Server instructions string
+ * Build server instructions for the given server mode.
  */
-export function getServerInstructions(uiMode?: UiMode): string {
-    return uiMode === 'openai'
-        ? getOpenaiInstructions()
-        : getDefaultInstructions();
+export function getServerInstructions(mode: ServerMode): string {
+    return instructionsByMode[mode]();
 }
